pyproject.toml
[tool.poetry]
name = "threadback"
version = "0.1.2"
description = ""
authors = ["Kazon Wilson <kazonwilson@gmail.com>"]

[tool.poetry.dependencies]
python = "3.8.3"
twint = "^2.1.20"
mongoengine = "^0.20.0"
pymongo = "^3.10.1"
uvicorn = "^0.11.5"
strawberry-graphql = "^0.27.2"
arrow = "^0.15.6"
huey = "^2.2.0"
redis = "^3.5.3"
twitter-scraper = {git = "https://github.com/bisguzar/twitter-scraper.git"}
dnspython = "^1.16.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pre-commit = "^2.5.1"
black = "^19.10b0"

[tool.poetry.scripts]
app = "run:start_dev"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
